renaming legacy talaiot
