+++ library/plugins/talaiot/src/main/kotlin/com/cdsap/talaiot/plugin/TalaiotPublisherConfiguration.kt
