+++ settings.gradle.kts
@@ -15,4 +15,4 @@ include(":library:talaiot")
 include(":library:talaiot-request")
 include(":library:talaiot-logger")
 include(":library:talaiot-test-utils")
-include(":library:plugins:talaiot-legacy")
+include(":library:plugins:talaiot")
